Bumped interplay version to 1.1.0 and dropped Scala 2.10
This is in preparation of releasing anorm against Play 2.5.0-M1

Because Play 2.5.0-M1 is only available for Scala 2.11, support for Scala 2.10
is dropped with this commit. The reason why we added an explicit `scalaVersion`
value in the `anorm-parent` project is because its  `scalaVersion` value was still
2.10.4. This is likely because something needs to be fixed in Play sbt interplay
plugin, which this build depend on.

Since we bumped the Play major version number, the plugin's version is also
bumped to 3.0.0.